+++ sdk/identity/Azure.Identity/src/ClientCertificateCredential.cs
@@ -32,7 +32,8 @@ namespace Azure.Identity
 
         internal IX509Certificate2Provider ClientCertificateProvider { get; }
 
-        private readonly MsalConfidentialClient _client;
+        internal MsalConfidentialClient Client { get; }
+
         private readonly CredentialPipeline _pipeline;
         private readonly bool _allowMultiTenantAuthentication;
 
@@ -131,7 +132,9 @@ namespace Azure.Identity
 
             _pipeline = pipeline ?? CredentialPipeline.GetInstance(options);
 
-            _client = client ?? new MsalConfidentialClient(_pipeline, tenantId, clientId, certificateProvider, (options as ClientCertificateCredentialOptions)?.SendCertificateChain ?? false, options as ITokenCacheOptions);
+            ClientCertificateCredentialOptions certCredOptions = (options as ClientCertificateCredentialOptions);
+
+            Client = client ?? new MsalConfidentialClient(_pipeline, tenantId, clientId, certificateProvider, certCredOptions?.SendCertificateChain ?? false, options as ITokenCacheOptions, certCredOptions?.RegionalAuthority);
         }
 
         /// <summary>
@@ -147,7 +150,7 @@ namespace Azure.Identity
             try
             {
                 var tenantId = TenantIdResolver.Resolve(TenantId, requestContext, _allowMultiTenantAuthentication);
-                AuthenticationResult result = _client.AcquireTokenForClientAsync(requestContext.Scopes, tenantId, false, cancellationToken).EnsureCompleted();
+                AuthenticationResult result = Client.AcquireTokenForClientAsync(requestContext.Scopes, tenantId, false, cancellationToken).EnsureCompleted();
 
                 return scope.Succeeded(new AccessToken(result.AccessToken, result.ExpiresOn));
             }
@@ -170,7 +173,7 @@ namespace Azure.Identity
             try
             {
                 var tenantId = TenantIdResolver.Resolve(TenantId, requestContext, _allowMultiTenantAuthentication);
-                AuthenticationResult result = await _client
+                AuthenticationResult result = await Client
                     .AcquireTokenForClientAsync(requestContext.Scopes, tenantId, true, cancellationToken)
                     .ConfigureAwait(false);