+++ sdk/webpubsub/Azure.Messaging.WebPubSub/src/WebPubSubAuthenticationPolicy.cs
@@ -14,6 +14,7 @@ namespace Azure.Messaging.WebPubSub
     internal partial class WebPubSubAuthenticationPolicy : HttpPipelineSynchronousPolicy
     {
         private readonly AzureKeyCredential _credential;
+        private volatile KeyBytesCache _keyCache = new KeyBytesCache(string.Empty); // it's volatile so that the cache update below is not reordered
 
         /// <summary>
         /// Creates an instance of the authentication policy
@@ -24,13 +25,23 @@ namespace Azure.Messaging.WebPubSub
         /// <inheritdoc/>
         public override void OnSendingRequest(HttpMessage message)
         {
-            string audience = message.Request.Uri.ToUri().AbsoluteUri;
+            string audience;
+            if (!TryGetAudience(message, out audience)) {
+                audience = message.Request.Uri.ToUri().AbsoluteUri;
+            }
+
             var now = DateTimeOffset.UtcNow;
             var expiresAt = now + TimeSpan.FromMinutes(5);
 
-            var keyBytes = Encoding.UTF8.GetBytes(_credential.Key);
+            var key = _credential.Key;
+            var cache = _keyCache;
+            if (!ReferenceEquals(key, cache.Key))
+            {
+                cache = new KeyBytesCache(key);
+                _keyCache = cache;
+            }
 
-            var writer = new JwtBuilder(keyBytes);
+            var writer = new JwtBuilder(cache.KeyBytes);
             writer.AddClaim(JwtBuilder.Nbf, now);
             writer.AddClaim(JwtBuilder.Exp, expiresAt);
             writer.AddClaim(JwtBuilder.Iat, now);
@@ -46,6 +57,36 @@ namespace Azure.Messaging.WebPubSub
             });
 
             message.Request.Headers.SetValue(HttpHeader.Names.Authorization, headerValue);
+        }
+
+        // this is to support API Management Server
+        private const string AUDIENCE_SETTING = nameof(WebPubSubAuthenticationPolicy) + ".Audience";
+        public static void SetAudience(HttpMessage message, Uri audience)
+        {
+            message.SetProperty(AUDIENCE_SETTING, audience.AbsoluteUri);
+        }
+
+        private static bool TryGetAudience(HttpMessage message, out string audience)
+        {
+            if (message.TryGetProperty(AUDIENCE_SETTING, out var jwtAudience) &&
+            	jwtAudience is string uri)
+            {
+            	audience = uri;
+                return true;
+            }
+            audience = default;
+            return false;
+        }
+
+        private sealed class KeyBytesCache
+        {
+            public KeyBytesCache(string key)
+            {
+                Key = key;
+                KeyBytes = Encoding.UTF8.GetBytes(key);
+            }
+            public readonly byte[] KeyBytes;
+            public readonly string Key;
         }
     }
 }